Fix invalid timing files silently breaking all replay
The current behavior when an invalid timing file is encountered in
either the list or replay mode causes the program to exit without useful
information, sans a line number and a brief error message. The proper
behavior in this error case is to provide the user an idea where the
error was encountered (in this case the file name, line number, and line
itself).