Fix unexpected O(N) behavior when replaying a single file
Previously when attempting to replay a single recording that was within
a directory containing a large number of other recordings, sudosh opened
and scanned through _every_ file within the directory before loading the
replay. This is due to shared code that called the count_dur function
within the main loop. This was fixed by scoping count_dur to conditions
where we need it (e.g. listing all replays).

This is a hacky fix that should really encourage the entire replay entry
function to be re-written as two seperate code paths to prevent we're
cases with O(N)'s or other broken files.