Set the append only flag on the timing, script, and record files
This commit sets the append only flag on the files so that a user cannot
truncate the files before exiting their shell.

After the files are done being written to, we set the immutable flag so
that they cannot be altered, we also chmod them to be readable by the
users primary group.